+++ tests/integration/routes/meetings/detail/-components/meeting-detail-header/component-test.ts
@@ -1,9 +1,10 @@
-import { render } from '@ember/test-helpers';
+import { render, settled } from '@ember/test-helpers';
 import setupMirage from 'ember-cli-mirage/test-support/setup-mirage';
 import { click } from 'ember-osf-web/tests/helpers';
 import { setupRenderingTest } from 'ember-qunit';
 import { TestContext } from 'ember-test-helpers';
 import hbs from 'htmlbars-inline-precompile';
+import moment from 'moment';
 import { module, test } from 'qunit';
 
 module('Integration | routes | meetings | detail | -components | meeting-detail-header', hooks => {
@@ -103,5 +104,79 @@ module('Integration | routes | meetings | detail | -components | meeting-detail-
         assert.dom('[data-test-meeting-format-body]').hasTextContaining(meeting.fieldNames.mail_message_body);
         assert.dom('[data-test-meeting-format-body]').hasTextContaining(meeting.fieldNames.mail_attachment);
         assert.dom('[data-test-meeting-panel-footer-note]').exists();
+    });
+
+    test('it handles location and dates properly', async function(assert) {
+        server.create('meeting', {
+            id: 'testmeeting',
+            name: 'Test Meeting',
+            location: undefined,
+            startDate: undefined,
+            endDate: undefined,
+        });
+        const meeting = await this.store.findRecord('meeting', 'testmeeting');
+        this.set('meeting', meeting);
+        await render(hbs`<Meetings::Detail::-Components::MeetingDetailHeader @meeting={{this.meeting}} />`);
+        assert.dom('[data-test-meeting-location-and-date]')
+            .hasText(
+                '',
+                'Nothing displayed when no location or dates defined',
+            );
+        const location = 'Timbuktu';
+        meeting.set('location', location);
+        await settled();
+        assert.dom('[data-test-meeting-location-and-date]')
+            .hasText(
+                location,
+                'Only location displayed when only location is defined',
+            );
+        const startDate = new Date('2000-01-02');
+        const formattedStartDate = moment(startDate).format('MMM DD, YYYY');
+        meeting.set('startDate', startDate);
+        await settled();
+        assert.dom('[data-test-meeting-location-and-date]')
+            .hasText(
+                `${location} | ${formattedStartDate} -`,
+                'Only location and start date displayed when only location and start date are defined',
+            );
+        const endDate = new Date('2000-01-03');
+        const formattedEndDate = moment(endDate).format('MMM DD, YYYY');
+        meeting.set('endDate', endDate);
+        meeting.set('startDate', undefined);
+        await settled();
+        assert.dom('[data-test-meeting-location-and-date]')
+            .hasText(
+                `${location} | - ${formattedEndDate}`,
+                'Only location and end date displayed when only location and end date are defined',
+            );
+        meeting.set('location', undefined);
+        await settled();
+        assert.dom('[data-test-meeting-location-and-date]')
+            .hasText(
+                `- ${formattedEndDate}`,
+                'Only end date displayed when only end date is defined',
+            );
+        meeting.set('startDate', startDate);
+        meeting.set('endDate', undefined);
+        await settled();
+        assert.dom('[data-test-meeting-location-and-date]')
+            .hasText(
+                `${formattedStartDate} -`,
+                'Only start date displayed when only start date is defined',
+            );
+        meeting.set('endDate', endDate);
+        await settled();
+        assert.dom('[data-test-meeting-location-and-date]')
+            .hasText(
+                `${formattedStartDate} - ${formattedEndDate}`,
+                'Only start date and end date displayed when only start date and end date are defined',
+            );
+        meeting.set('location', location);
+        await settled();
+        assert.dom('[data-test-meeting-location-and-date]')
+            .hasText(
+                `${location} | ${formattedStartDate} - ${formattedEndDate}`,
+                'Location and dates displayed when all are defined',
+            );
     });
 });